M2-5960: moved Feedback utils to separate files to avoid circular dependencies
