+++ src/modules/Dashboard/features/RespondentData/RespondentDataReview/Feedback/FeedbackReviews/FeedbackReviews.test.tsx
@@ -22,7 +22,7 @@ import { FeedbackReviews } from './FeedbackReviews';
 import { AssessmentActivityItem } from '../../RespondentDataReview.types';
 import { RespondentDataReviewContext } from '../../RespondentDataReview.context';
 import { FeedbackForm } from '../Feedback.types';
-import { getDefaultFormValues } from '../Feedback.utils';
+import { getDefaultFormValues } from '../utils/getDefaultValues';
 
 const mockedAnswerId = '0a7bcd14-24a3-48ed-8d6b-b059a6541ae4';
 const route = `/dashboard/${mockedAppletId}/respondents/${mockedRespondent}/dataviz/responses?selectedDate=2023-11-27&answerId=${mockedAnswerId}`;
